requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multi-key-dict==2.0.3
pbr==5.4.3
python-jenkins==1.5.0
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.6
